LayoutTests/http/tests/websocket/tests/hybi/handshake-ok-with-legacy-sec-websocket-response-headers-expected.txt

Test whether WebSocket handshake is OK with legacy response headers 'Sec-WebSocket-Location' and 'Sec-WebSocket-Origin'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.isTrusted is true
PASS successfullyParsed is true

TEST COMPLETE
